fix: handle macos terminal pulse renames
